requirements.txt
python-telegram-bot==20.7
groq==0.4.1
python-dotenv==1.0.0